2019-01-28 Relationships and Routes Protection

+++ controllers/videoController.js
@@ -66,13 +66,17 @@ export const videoDetail = async (req, res) => {
 	}
 };
 
-export const getEditVideo = async(req, res) => {
+export const getEditVideo = async (req, res) => {
 	const {
 		params: {id}
 	} = req;
 	try {
 		const video = await Video.findById(id);
-		res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+		}
 	} catch (error) {
 		res.redirect(routes.home);
 	}
@@ -94,11 +98,17 @@ export const postEditVideo = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
 	const {
-		params: {id}
+		params: { id }
 	} = req;
-
 	try {
-		await Video.findOneAndDelete({_id: id});
-	} catch (error) {};
+		const video = await Video.findById(id);
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			await Video.findOneAndDelete({ _id: id });
+		}
+	} catch (error) {
+		console.log(error);
+	}
 	res.redirect(routes.home);
 };
